api/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
python-dotenv
pymysql
numpy
scikit-learn
python-multipart
elasticsearch==9.0.2
